# Api-get-way/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/user_service/kitchen/create": {
            "post": {
                "description": "Create a new kitchen based on the provided request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kitchen"
                ],
                "summary": "Create a new kitchen",
                "parameters": [
                    {
                        "description": "Kitchen details to create",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.CreateKitchenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.CreateKitchenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/get_all": {
            "get": {
                "description": "Retrieve a list of all kitchens with optional pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kitchen"
                ],
                "summary": "Get all kitchens",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.KitchensResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/search": {
            "get": {
                "description": "Search for kitchens based on various criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kitchen"
                ],
                "summary": "Search kitchens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the kitchen to search for",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Rating of the kitchen",
                        "name": "rating",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Address of the kitchen",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Total orders for the kitchen",
                        "name": "total_orders",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Phone number of the kitchen",
                        "name": "phone_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cuisine type of the kitchen",
                        "name": "cuisine_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Description of the kitchen",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Owner ID of the kitchen",
                        "name": "owner_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.KitchensResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/update-working-hours": {
            "put": {
                "description": "Updates the working hours of a kitchen based on the provided request payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Update kitchen working hours",
                "operationId": "updateWorkingHours",
                "parameters": [
                    {
                        "description": "Request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateWorkingHoursRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated working hours information",
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateWorkingHoursResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/{id}": {
            "get": {
                "description": "Retrieve a kitchen by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kitchen"
                ],
                "summary": "Get a kitchen by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID to fetch",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.KitchenResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/{kitchen_id}": {
            "put": {
                "description": "Update an existing kitchen based on the provided request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kitchen"
                ],
                "summary": "Update an existing kitchen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID to update",
                        "name": "kitchen_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated kitchen details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateKitchenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateKitchenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/{kitchen_id}/statistics": {
            "get": {
                "description": "Retrieves statistics for a kitchen based on the provided kitchen_id, start_date, and end_date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Retrieve kitchen statistics",
                "operationId": "getKitchenStatistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID",
                        "name": "kitchen_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics for the kitchen",
                        "schema": {
                            "$ref": "#/definitions/genproto.KitchenStatisticsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/users/activity": {
            "get": {
                "description": "Retrieves activity statistics for a user based on the provided start_date and end_date query parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Retrieve user activity",
                "operationId": "getUserActivity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity statistics for the user",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetUserActivityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/users/{id}/profile": {
            "get": {
                "description": "Retrieves a user's profile information based on the provided ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserProfile"
                ],
                "summary": "Get user profile by ID",
                "operationId": "get-user-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a user's profile information based on the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserProfile"
                ],
                "summary": "Update user profile",
                "operationId": "update-user-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User Profile Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "genproto.CreateKitchenRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "genproto.CreateKitchenResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
...
        "genproto.GetUserActivityResponse": {
            "type": "object",
            "properties": {
                "favorite_cuisines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.CuisineActivity"
                    }
                },
                "favorite_kitchens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.KitchenActivity"
                    }
                },
                "total_orders": {
                    "type": "integer"
                },
                "total_spent": {
                    "type": "number"
                }
            }
        },
...
        "genproto.KitchenResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "total_order": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "genproto.KitchensResponse": {
            "type": "object",
            "properties": {
                "kitchens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.KitchenResponse"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "genproto.UpdateKitchenRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "genproto.UpdateKitchenResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "genproto.UpdateWorkingHoursRequest": {
            "type": "object",
            "properties": {
                "kitchen_id": {
                    "type": "string"
                },
                "working_hours": {
                    "$ref": "#/definitions/genproto.WorkingHours"
                }
            }
        },
        "genproto.UpdateWorkingHoursResponse": {
            "type": "object",
            "properties": {
                "kitchen_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "working_hours": {
                    "$ref": "#/definitions/genproto.WorkingHours"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}